python/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
python-multipart==0.0.6 